Added IsDeletable and MarkForDelete to IRelationship and its relevant Children Changed BusinessObject IsDeletable to use its Relationships IsDeletable so that if the BusinessObject is not deletable due to a rule from its relationship the BusinessObject.IsDeletable will now return false. Modified the BusinessObject.MarkForDelete to call the MarkForDelete of all of its Relationships. The Relationships will therefore mark all related objects that are related via a relationshionship that is DeleteRelated.

+++ source/Habanero.Test.Bo/TestPack/ContactPersonTestBO.cs
@@ -704,7 +704,7 @@ namespace Habanero.Test.BO
 					<primaryKey>
 						<prop name=""ContactPersonID"" />
 					</primaryKey>
-                    <relationship name=""Organisation"" type=""single"" relatedClass=""OrganisationTestBO"" relatedAssembly=""Habanero.Test.BO"" reverseRelationship=""ContactPeople"">
+                    <relationship name=""Organisation"" type=""single"" relatedClass=""OrganisationTestBO"" relatedAssembly=""Habanero.Test.BO"" reverseRelationship=""ContactPeople"" deleteAction=""DoNothing"">
 						<relatedProperty property=""OrganisationID"" relatedProperty=""OrganisationID"" />
 					</relationship>
 					<ui>
@@ -735,7 +735,7 @@ namespace Habanero.Test.BO
 					<primaryKey>
 						<prop name=""ContactPersonID"" />
 					</primaryKey>
-                    <relationship name=""Organisation"" type=""single"" relatedClass=""OrganisationTestBO"" relatedAssembly=""Habanero.Test.BO"" reverseRelationship=""ContactPerson"">
+                    <relationship name=""Organisation"" type=""single"" relatedClass=""OrganisationTestBO"" relatedAssembly=""Habanero.Test.BO"" reverseRelationship=""ContactPerson"" deleteAction=""DoNothing"">
 						<relatedProperty property=""OrganisationID"" relatedProperty=""OrganisationID"" />
 					</relationship>
 					<ui>
@@ -885,6 +885,10 @@ namespace Habanero.Test.BO
 
         public static void DeleteAllContactPeople()
         {
+            if (BORegistry.DataAccessor is DataAccessorInMemory)
+            {
+                return;
+            }
             string sql = "DELETE FROM contact_person_address";
             DatabaseConnection.CurrentConnection.ExecuteRawSql(sql);
             sql = "DELETE FROM Contact_Person";
